refactor(llm): move openai tool_search_passthrough into model_constructors
