+++ front/lib/model_constructors/sdk/openai_responses/converters/input/utils.ts
@@ -1,5 +1,5 @@
 import { OPENAI_TOOL_SEARCH_TOOL } from "@app/lib/api/llm/clients/openai/utils/tool_search";
-import { parseOpenAIToolSearchItem } from "@app/lib/api/llm/clients/openai/utils/tool_search_passthrough";
+import { parseOpenAIToolSearchItem } from "@app/lib/model_constructors/sdk/openai_responses/converters/input/tool_search_passthrough";
 import type {
   OutputFormat,
   Reasoning,
